requirements.txt
# MedClaw 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pandas==2.1.4
numpy==1.24.3
pyyaml==6.0.1
httpx==0.25.2

# 系统监控
psutil==5.9.6

# Computer-Use 能力
pyautogui==0.9.54
pytesseract==0.3.10
pillow==10.1.0
opencv-python==4.8.1.78
pygetwindow==0.0.9
pyperclip==1.8.2
pytesseract==0.3.10

# MCP 客户端
# mcp-client==0.2.0  # 如果有此库则安装

# QQBot - NoneBot2 (二选一)
nonebot2==2.2.0
nonebot-adapter-qq==1.4.0

# QQBot - go-cqhttp (二选一，如使用 go-cqhttp 则不需要 nonebot)
websocket-client==1.7.0

# 病历质控模块
fuzzywuzzy==0.18.0
python-dateutil==2.8.2

# 医保分析模块
# drg-tools==0.1.0  # 假设有此库

# 临床预测模块
scikit-learn==1.3.2
joblib==1.3.2

# AI 阅片模块
# torch==2.1.0  # 根据需要选择，CUDA 版本需匹配
# torchvision==0.16.0
pydicom==2.4.4

# 论文分析模块
pdfplumber==0.10.3

# 办公自动化
# selenium==4.15.2  # 网页自动化需要

# 其他工具
python-multipart==0.0.6
aiofiles==23.2.1
